final app

+++ amazed/controllers/detailsController.js
@@ -4,35 +4,29 @@ import Review from '../model/reviewModel.js';
 
 //Add product detail
 export const addNew = async (req, res) => {
-    const { asin, title, description, price, discount, images, dimensions, weight, manufacturer, model_number, sold_by, brand } = req.body;
-
+    const { asin, title, description,reviews, price, discount, images, dimensions, weight, manufacturer, model_number, sold_by, brand } = req.body;
+    const features = req.body.features.split(',');
     const date = new Date();
     const month = date.toLocaleString('default', { month: 'long' });
 
     try {
+        if(!req.session.user && req.session.user.role !=='Admin') {	
+            return res.status(400).send('No Session Found! Please Login Again')	
+        }
+        
         let data = {
             asin: asin,
             title: title,
             description: description,
-            feature_bullets: req.body.features.split(','),
-            reviews: {
-                rating: "0.0",
-            },
-            price: {
-                symbol: "₹",
-                currency: "INR",
-                current_price: (price - discount),
-                discounted: discount?true:false,
-                before_price: price,
-                savings_amount: discount,
-                savings_percent: (discount/price)*100
-            },
+            feature_bullets: features,
+            reviews: reviews,
+            price: price,
             total_images: images.length,
             images: images,
             product_information: {
                 dimensions: dimensions,
                 weight: weight,
-                available_from: `${date.getDay()} ${month} ${date.getFullYear()}`,
+                available_from: `${date.getDate()} ${month} ${date.getFullYear()}`,
                 available_from_utc: date.toISOString(),
                 manufacturer: manufacturer,
                 model_number: model_number,
@@ -77,13 +71,15 @@ export const getDetail = async (req, res) => {
 }
 
 //Delete detail
-export const deleteDetail = async (req, res) => {
-    try {
-        const data = await Details.deleteOne({"asin": req.body.asin});
-
-        return res.status(200).send({"success": "Deleted Successfully!"})
-    }
-    catch(error) {
-        return res.status(409).send({"err": error.message});
-    }
+export const deleteDetail = async (req, res) => {	
+    try {	
+        if(!req.session.user && req.session.user.role !=='Admin') {	
+            return res.status(400).send('No Session Found! Please Login Again')	
+        }	
+        const data = await Details.deleteOne({"asin": req.body.asin});	
+        return res.status(200).send({"success": "Deleted Successfully!"})	
+    }	
+    catch(error) {	
+        return res.status(409).send({"err": error.message});	
+    }	
 }
