+++ amazed/controllers/ordersController.js
@@ -1,8 +1,44 @@
 import Orders from '../model/ordersModel.js';
+import { checkEmail, checkNumber, checkString } from '../utils/validator.js';
 
 //Place order
 export const newOrder = async(req,res) => {
+    
     try{
+
+        // if(!req.session.user && req.session.user.role !== 'User') {
+        //     return res.status(400).send('No Session Found! Please Login Again')
+        // }
+
+        // if((req.session.user.name !== req.body.userName) || (req.session.user.email !== req.body.userEmail)) {
+        //     return res.status(400).send('No Session Found! Please Login Again')
+        // }
+
+        // const IsValidUserEmail = checkEmail(req.body.userEmail);
+        // if(!IsValidUserEmail){
+        //     res.send("Invalid Email")
+        // }
+
+        // const IsValidEmail = checkEmail(req.body.email);
+        // if(!IsValidEmail){
+        //     res.send("Invalid Email")
+        // }
+
+        // const IsValidFName = checkString(req.body.fname);
+        // if(!IsValidFName){
+        //     res.send("Invalid first name")
+        // }
+
+        // const IsValidLName = checkString(req.body.lname);
+        // if(!IsValidLName){
+        //     res.send("Invalid last name")
+        // }
+
+        // const IsValidPhone = checkNumber(req.body.phone);
+        // if(!IsValidPhone){
+        //     res.send("Invalid phone number")
+        // }
+        console.log("payment", req.body.payment)
         let data = {
             date : req.body.date,
             userName : req.body.userName,
@@ -10,14 +46,17 @@ export const newOrder = async(req,res) => {
             orderDetails : {
                 firstName : req.body.fname,
                 lastName : req.body.lname,
-                houseAddress : req.body.houseadd,
-                apartment : req.body.apartment,
+                houseAddress : req.body.houseadd ? req.body.houseadd: '',
+                apartment : req.body.apartment ? req.body.apartment: '',
                 city : req.body.city,
                 state : req.body.state,
                 country : req.body.country,
                 postCode : req.body.postCode,
                 phone : req.body.phone,
-                email : req.body.email
+                email : req.body.email,
+                paymentMode : req.body.payment,
+                transactionId:req.body.transactionid,
+                transactionAmount:req.body.transactionamount
             },
             productDetails : {
                 asin : req.body.asin,
@@ -27,7 +66,9 @@ export const newOrder = async(req,res) => {
                 currentPrice : req.body.currentPrice,
                 quantity : req.body.quantity,
                 totalPrice : req.body.totalPrice,
-                paymentMode : req.body.payment
+                couponDiscount: req.body.couponDiscount,
+                grandTotal: req.body.grandTotal
+                
             },
             orderStatus : req.body.status,
             delivered : req.body.delivered,
@@ -47,10 +88,14 @@ export const newOrder = async(req,res) => {
         }
 };
 
-//get orders
+//get All orders
 export const allOrders = async(req,res) => {
     try{
-        const result = await Orders.find({})
+        if(!req.session.user && req.session.user.role !=='Admin') {
+            return res.status(400).send('No Session Found! Please Login Again')
+        }
+
+        const result = await Orders.find({}).sort({date: -1})
 
         if(result.length <1) return res.status(404).send({"err":"No Data Found"});
         res.status(200).send(result)
@@ -63,10 +108,12 @@ export const allOrders = async(req,res) => {
 
 // get order by id
 export const ordersById = async(req,res) => {
-    const _id = req.params.id
-    console.log(_id)
     try{
         
+        if(!req.session.user) {
+            return res.status(400).send('No Session Found! Please Login Again')
+        }
+
         const result = await Orders.findById(req.params.id)
         if(result.length <1) return res.status(404).send({"err":"No Data Found"});
         res.status(200).send(result)
@@ -81,6 +128,11 @@ export const ordersById = async(req,res) => {
 //get orders by userEmail
 export const ordersByEmail = async(req,res) => {
     try{
+
+        // if(!req.session.user) {
+        //     return res.status(400).send('No Session Found! Please Login Again')
+        // }
+
         const userEmail = req.query.userEmail
         const result = await Orders.find({userEmail:userEmail}).sort({ date: -1})
 
@@ -96,16 +148,24 @@ export const ordersByEmail = async(req,res) => {
 //update order
 export const updateOrder = async (req, res) => {
     const user = req.body;
-
+    
     try {
-        const _id = req.params.id
-        const updaterequired = await Orders.findByIdAndUpdate(_id, {orderStatus:req.body.status});
 
-        res.status(204).send({"sucess":"Order is updated successfully"});
+        if(!req.session.user) {
+            return res.status(400).send('No Session Found! Please Login Again')
+        }
+
+        const _id = req.params.id
+        await Orders.findByIdAndUpdate(_id, {
+            "delivered":req.body.delivered, 
+            "orderStatus":req.body.orderStatus
+        });
+
+        return res.status(200).send({"success":"Order is updated successfully"});
     }
 
     catch (error) {
-        res.status(404).send({"err":error.message})
+        return res.status(404).send({"err":error.message})
     }
 };
 
@@ -113,6 +173,11 @@ export const updateOrder = async (req, res) => {
 export const deleteOrder = async (req, res) => {
 
     try {
+
+        if(!req.session.user) {
+            return res.status(400).send('No Session Found! Please Login Again')
+        }
+
         const _id = req.params.id
         const deleted = await Orders.findByIdAndDelete(_id);
